Refactor based on PR (#286)
