+++ apps/coscrad-frontend/src/components/resources/media-items/media-item-index.presenter.tsx
@@ -3,7 +3,7 @@ import { MediaItemIndexState } from '../../../store/slices/resources/media-items
 import { HeadingLabel, IndexTable } from '../../../utils/generic-components/presenters/tables';
 import { CellRenderersDefinition } from '../../../utils/generic-components/presenters/tables/generic-index-table-presenter/types/cell-renderers-definition';
 import { renderAggregateIdCell } from '../utils/render-aggregate-id-cell';
-import { formatBilingualText } from '../vocabulary-lists/utils/formatBilingualText';
+import { formatBilingualText } from '../vocabulary-lists/utils';
 
 export const MediaItemIndexPresenter = ({ entities: mediaItems }: MediaItemIndexState) => {
     const headingLabels: HeadingLabel<IMediaItemViewModel>[] = [
