+++ apps/coscrad-frontend/src/components/resources/vocabulary-lists/utils/format-bilingual-text.ts
@@ -1,6 +1,5 @@
 import { isEmptyText } from './isEmptyText';
 
-// TODO Add unit test
 export const formatBilingualText = (
     textInPrimaryLanguage: string,
     textInSecondaryLanguage: string
