+++ apps/coscrad-frontend/src/components/resources/vocabulary-lists/utils/isEmptyText.ts
@@ -1,5 +1,5 @@
 import { isNullOrUndefined } from '@coscrad/validation-constraints';
 
-// TODO Move to `@coscrad/validation-constraints`
+// TODO Consider moving this to `@coscrad/validation-constraints`
 export const isEmptyText = (input: string | null | undefined): boolean =>
     isNullOrUndefined(input) || input === '';
