--- apps/coscrad-frontend/src/configurable-front-matter/data/content.config.SAMPLE.json
@@ -1,15 +1,0 @@
-{
-    "siteTitle": "Site Title",
-    "subTitle": "Site Subtitle",
-    "about": "Add a few sentences about your web of knowledge!",
-    "siteDescription": "Your site decription goes here..",
-    "siteHomeImageUrl": "https://www.picbox.org/home-image.png",
-    "copyrightHolder": "My Organization",
-    "organizationLogoUrl": "https://www.picbox.org/logo.png",
-    "songIdToCredits": {
-        "1": "Credits for song 1"
-    },
-    "videoIdToCredits": {
-        "2": "Johnny produced this one!"
-    }
-}
