+++ libs/validation-constraints/src/lib/__tests__/constants.ts
@@ -1,1 +1,1 @@
-export const IT_SHOULD_RETURN_THE_EXPECTED_RESULT = 'it should return the expected result';
+export const SHOULD_RETURN_THE_EXPECTED_RESULT = 'should return the expected result';
