+++ libs/validation-constraints/src/lib/constraints/is-function.ts
@@ -1,2 +1,4 @@
-export const isFunction = (input: unknown): input is (args: unknown) => unknown =>
-    typeof input === 'function';
+// This might be useful in a utility types lib
+type UnknownFunction = (args: unknown) => unknown;
+
+export const isFunction = (input: unknown): input is UnknownFunction => typeof input === 'function';
