+++ libs/validation-constraints/src/lib/constraints/is-integer.spec.ts
@@ -3,7 +3,7 @@ import {
     assertConstraintSatisfiedForPredicate,
     buildInvalidCaseDescription,
     buildValidCaseDescription,
-    IT_SHOULD_RETURN_THE_EXPECTED_RESULT,
+    SHOULD_RETURN_THE_EXPECTED_RESULT,
 } from '../__tests__';
 import { isInteger } from './is-integer';
 
@@ -32,7 +32,7 @@ describe('isInteger', () => {
 
         validValues.forEach((validValue) => {
             describe(buildValidCaseDescription(validValue), () => {
-                it(IT_SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
+                it(SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
                     assertConstraintSatisfied(validValue);
                 });
             });
@@ -44,7 +44,7 @@ describe('isInteger', () => {
 
         invalidValues.forEach((invalidValue) => {
             describe(buildInvalidCaseDescription(invalidValue), () => {
-                it(IT_SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
+                it(SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
                     assertConstraintFailure(invalidValue);
                 });
             });
