+++ libs/validation-constraints/src/lib/constraints/is-non-negative-finite-number.spec.ts
@@ -3,7 +3,7 @@ import {
     assertConstraintSatisfiedForPredicate,
     buildInvalidCaseDescription,
     buildValidCaseDescription,
-    IT_SHOULD_RETURN_THE_EXPECTED_RESULT,
+    SHOULD_RETURN_THE_EXPECTED_RESULT,
 } from '../__tests__';
 import { isNonNegativeFiniteNumber } from './is-non-negative-finite-number';
 
@@ -29,7 +29,7 @@ describe('isNonNegativeFiniteNumber', () => {
 
         validValues.forEach((validValue) => {
             describe(buildValidCaseDescription(validValue), () => {
-                it(IT_SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
+                it(SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
                     assertConstraintSatisfied(validValue);
                 });
             });
@@ -42,7 +42,7 @@ describe('isNonNegativeFiniteNumber', () => {
 
         invalidValues.forEach((invalidValue) => {
             describe(buildInvalidCaseDescription(invalidValue), () => {
-                it(IT_SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
+                it(SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
                     assertConstraintFailure(invalidValue);
                 });
             });
