+++ libs/validation-constraints/src/lib/constraints/is-null.spec.ts
@@ -4,7 +4,7 @@ import {
     assertConstraintSatisfiedForPredicate,
     buildInvalidCaseDescription,
     buildValidCaseDescription,
-    IT_SHOULD_RETURN_THE_EXPECTED_RESULT,
+    SHOULD_RETURN_THE_EXPECTED_RESULT,
 } from '../__tests__';
 
 // This is the only value that satisfies the constraint `isNull`
@@ -33,7 +33,7 @@ describe('isNull', () => {
         const assertConstraintSatisfied = assertConstraintSatisfiedForPredicate(isNull);
 
         describe(buildValidCaseDescription(validValue), () => {
-            it(IT_SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
+            it(SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
                 assertConstraintSatisfied(validValue);
             });
         });
@@ -44,7 +44,7 @@ describe('isNull', () => {
 
         invalidValues.forEach((invalidValue) => {
             describe(buildInvalidCaseDescription(invalidValue), () => {
-                it(IT_SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
+                it(SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
                     assertConstraintFailure(invalidValue);
                 });
             });
