+++ libs/validation-constraints/src/lib/constraints/is-year.spec.ts
@@ -3,7 +3,7 @@ import {
     assertConstraintSatisfiedForPredicate,
     buildInvalidCaseDescription,
     buildValidCaseDescription,
-    IT_SHOULD_RETURN_THE_EXPECTED_RESULT,
+    SHOULD_RETURN_THE_EXPECTED_RESULT,
 } from '../__tests__';
 import { isYear } from './is-year';
 
@@ -27,7 +27,7 @@ describe('isYear', () => {
 
         validValues.forEach((validValue) => {
             describe(buildValidCaseDescription(validValue), () => {
-                it(IT_SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
+                it(SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
                     assertConstraintSatisfied(validValue);
                 });
             });
@@ -38,7 +38,7 @@ describe('isYear', () => {
 
         invalidValues.forEach((invalidValue) => {
             describe(buildInvalidCaseDescription(invalidValue), () => {
-                it(IT_SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
+                it(SHOULD_RETURN_THE_EXPECTED_RESULT, () => {
                     assertConstraintFailure(invalidValue);
                 });
             });
