--- libs/validation-constraints/src/lib/validation-constraints.ts
@@ -1,3 +1,0 @@
-export function validationConstraints(): string {
-    return 'validation-constraints';
-}
